PMA-156 - updated Parking pass resource

+++ database/migrations/2019_10_13_164233_create_parking_pass_logs_table.php
@@ -14,10 +14,10 @@ class CreateParkingPassLogsTable extends Migration
     public function up()
     {
         Schema::create('parking_pass_logs', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->integer('createdByUserId')->unsigned()->nullable();
             $table->unsignedInteger('propertyId');
-            $table->unsignedInteger('passId');
+            $table->unsignedBigInteger('passId');
             $table->unsignedInteger('spaceId');
             $table->string('make', 100)->nullable();
             $table->string('model', 100)->nullable();
